docs(cms-plugin): document brand root path model
