+++ libs/cms-plugin/src/collections/brands/config.ts
@@ -1,30 +1,20 @@
-import type { TFunction } from "@payloadcms/translations";
 import type {
   CollectionConfig,
   Locale,
   SanitizedCollectionConfig,
-  TextField,
-  ValidateOptions,
 } from "payload";
 
-import { text } from "payload/shared";
-
 import type { RowLabelProps } from "../../components/client/row-label.js";
-import type { TranslationsKey } from "../../translations/types.js";
 
 import { canManageContent, isAdmin } from "../../common/access-control.js";
 import { getLivePreviewUrl } from "../../common/live-preview.js";
-import {
-  normalizePathnameInput,
-  validateRootPathFormat,
-} from "../../common/pathname.js";
 import { imageField } from "../../fields/image.js";
 import { linkField } from "../../fields/link.js";
 import { showField } from "../../fields/show.js";
 import { textField } from "../../fields/text.js";
 import { contentGroup } from "../../groups.js";
 import { syncBrandHomeLink } from "./home-link.js";
-import { getBrandsForRootPath } from "./root-path.js";
+import { rootPathField } from "./root-path.js";
 import { brandUsagesField } from "./usages.js";
 
 type BrandsOptions = {
@@ -126,67 +116,7 @@ export function Brands({
                 },
                 required: false,
               }),
-              textField({
-                name: "rootPath",
-                admin: {
-                  description: {
-                    en: "The localized root path for this brand. Pages for this brand must use this path or a child path below it.",
-                    es: "La ruta raíz localizada de esta marca. Las páginas de esta marca deben usar esta ruta o una ruta hija dentro de ella.",
-                  },
-                  placeholder: "e.g. /brand-name",
-                },
-                hooks: {
-                  beforeValidate: [
-                    ({ value }) =>
-                      typeof value === "string"
-                        ? normalizePathnameInput(value)
-                        : value,
-                  ],
-                },
-                label: {
-                  en: "Root Path",
-                  es: "Ruta raíz",
-                },
-                validate: async (
-                  value: null | string | undefined,
-                  options: ValidateOptions<
-                    Record<string, unknown>,
-                    Record<string, unknown>,
-                    TextField,
-                    string
-                  >,
-                ) => {
-                  const defaultValidationResult = text(value, options);
-                  if (defaultValidationResult !== true) {
-                    return defaultValidationResult;
-                  }
-
-                  const t = options.req
-                    .t as unknown as TFunction<TranslationsKey>;
-                  const rootPath = normalizePathnameInput(value ?? "");
-                  const formatValidationResult =
-                    validateRootPathFormat(rootPath);
-
-                  if (formatValidationResult !== true) {
-                    return t(
-                      `cmsPlugin:brands:rootPath:${formatValidationResult}`,
-                    );
-                  }
-
-                  const brands = await getBrandsForRootPath(
-                    options.req,
-                    rootPath,
-                  );
-                  const alreadyExists = brands.some(
-                    (brand) => brand.id !== options.id,
-                  );
-                  if (alreadyExists) {
-                    return t("cmsPlugin:brands:rootPath:alreadyExists");
-                  }
-
-                  return true;
-                },
-              }),
+              rootPathField(),
               textField({
                 name: "baseTitle",
                 admin: {
